optimization: avoid GetTrail when meter specified

+++ stepmania/src/Course.h
@@ -147,7 +147,7 @@ public:
 	Trail* GetTrailForceRegenCache( StepsType st, CourseDifficulty cd=DIFFICULTY_MEDIUM ) const;
 	void GetTrails( vector<Trail*> &AddTo, StepsType st ) const;
 	void GetAllTrails( vector<Trail*> &AddTo ) const;
-	float GetMeter( StepsType st, CourseDifficulty cd=DIFFICULTY_MEDIUM ) const;
+	int GetMeter( StepsType st, CourseDifficulty cd=DIFFICULTY_MEDIUM ) const;
 	bool HasMods() const;
 	bool AllSongsAreFixed() const;
